Moved Plugins from plugin package to module.
