+++ spring-xd-module/src/main/java/org/springframework/xd/module/AbstractPlugin.java
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package org.springframework.xd.plugin;
+package org.springframework.xd.module;
 
 
 
@@ -23,7 +23,6 @@ import java.util.List;
 import org.springframework.context.ConfigurableApplicationContext;
 import org.springframework.core.io.ClassPathResource;
 import org.springframework.util.Assert;
-import org.springframework.xd.module.Module;
 
 /*
  * @author Glenn Renfro
