+++ spring-xd-module/src/main/java/org/springframework/xd/module/BeanDefinitionAddingPostProcessor.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.springframework.xd.plugin;
+package org.springframework.xd.module;
 
 import java.util.Map.Entry;
 import java.util.SortedMap;
